Split tests for aminotypes and move to modules
