+++ packages/stargate/src/modules/staking/aminomessages.spec.ts
@@ -1,0 +1,351 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+import { encodeBech32Pubkey } from "@cosmjs/amino";
+import { fromBase64 } from "@cosmjs/encoding";
+import { coin } from "@cosmjs/proto-signing";
+import {
+  MsgBeginRedelegate,
+  MsgCreateValidator,
+  MsgDelegate,
+  MsgEditValidator,
+  MsgUndelegate,
+} from "cosmjs-types/cosmos/staking/v1beta1/tx";
+
+import { AminoTypes } from "../../aminotypes";
+import {
+  AminoMsgBeginRedelegate,
+  AminoMsgCreateValidator,
+  AminoMsgDelegate,
+  AminoMsgEditValidator,
+  AminoMsgUndelegate,
+  createStakingAminoConverters,
+} from "./aminomessages";
+
+describe("AminoTypes", () => {
+  describe("toAmino", () => {
+    it("works for MsgBeginRedelegate", () => {
+      const msg: MsgBeginRedelegate = {
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorSrcAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        validatorDstAddress: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
+        amount: coin(1234, "ucosm"),
+      };
+      const aminoTypes = new AminoTypes({
+        ...createStakingAminoConverters("cosmos"),
+      });
+      const aminoMsg = aminoTypes.toAmino({
+        typeUrl: "/cosmos.staking.v1beta1.MsgBeginRedelegate",
+        value: msg,
+      });
+      const expected: AminoMsgBeginRedelegate = {
+        type: "cosmos-sdk/MsgBeginRedelegate",
+        value: {
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_src_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          validator_dst_address: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
+          amount: coin(1234, "ucosm"),
+        },
+      };
+      expect(aminoMsg).toEqual(expected);
+    });
+
+    it("works for MsgCreateValidator", () => {
+      const msg: MsgCreateValidator = {
+        description: {
+          moniker: "validator",
+          identity: "me",
+          website: "valid.com",
+          securityContact: "Hamburglar",
+          details: "...",
+        },
+        commission: {
+          rate: "0.2",
+          maxRate: "0.3",
+          maxChangeRate: "0.1",
+        },
+        minSelfDelegation: "123",
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        pubkey: {
+          typeUrl: "/cosmos.crypto.secp256k1.PubKey",
+          value: fromBase64("A08EGB7ro1ORuFhjOnZcSgwYlpe0DSFjVNUIkNNQxwKQ"),
+        },
+        value: coin(1234, "ucosm"),
+      };
+      const aminoTypes = new AminoTypes({
+        ...createStakingAminoConverters("cosmos"),
+      });
+      const aminoMsg = aminoTypes.toAmino({
+        typeUrl: "/cosmos.staking.v1beta1.MsgCreateValidator",
+        value: msg,
+      });
+      const expected: AminoMsgCreateValidator = {
+        type: "cosmos-sdk/MsgCreateValidator",
+        value: {
+          description: {
+            moniker: "validator",
+            identity: "me",
+            website: "valid.com",
+            security_contact: "Hamburglar",
+            details: "...",
+          },
+          commission: {
+            rate: "0.2",
+            max_rate: "0.3",
+            max_change_rate: "0.1",
+          },
+          min_self_delegation: "123",
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          pubkey: encodeBech32Pubkey(
+            { type: "tendermint/PubKeySecp256k1", value: "A08EGB7ro1ORuFhjOnZcSgwYlpe0DSFjVNUIkNNQxwKQ" },
+            "cosmos",
+          ),
+          value: coin(1234, "ucosm"),
+        },
+      };
+      expect(aminoMsg).toEqual(expected);
+    });
+
+    it("works for MsgDelegate", () => {
+      const msg: MsgDelegate = {
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        amount: coin(1234, "ucosm"),
+      };
+      const aminoTypes = new AminoTypes({
+        ...createStakingAminoConverters("cosmos"),
+      });
+      const aminoMsg = aminoTypes.toAmino({
+        typeUrl: "/cosmos.staking.v1beta1.MsgDelegate",
+        value: msg,
+      });
+      const expected: AminoMsgDelegate = {
+        type: "cosmos-sdk/MsgDelegate",
+        value: {
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          amount: coin(1234, "ucosm"),
+        },
+      };
+      expect(aminoMsg).toEqual(expected);
+    });
+
+    it("works for MsgEditValidator", () => {
+      const msg: MsgEditValidator = {
+        description: {
+          moniker: "validator",
+          identity: "me",
+          website: "valid.com",
+          securityContact: "Hamburglar",
+          details: "...",
+        },
+        commissionRate: "0.2",
+        minSelfDelegation: "123",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+      };
+      const aminoTypes = new AminoTypes({
+        ...createStakingAminoConverters("cosmos"),
+      });
+      const aminoMsg = aminoTypes.toAmino({
+        typeUrl: "/cosmos.staking.v1beta1.MsgEditValidator",
+        value: msg,
+      });
+      const expected: AminoMsgEditValidator = {
+        type: "cosmos-sdk/MsgEditValidator",
+        value: {
+          description: {
+            moniker: "validator",
+            identity: "me",
+            website: "valid.com",
+            security_contact: "Hamburglar",
+            details: "...",
+          },
+          commission_rate: "0.2",
+          min_self_delegation: "123",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        },
+      };
+      expect(aminoMsg).toEqual(expected);
+    });
+
+    it("works for MsgUndelegate", () => {
+      const msg: MsgUndelegate = {
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        amount: coin(1234, "ucosm"),
+      };
+      const aminoTypes = new AminoTypes({
+        ...createStakingAminoConverters("cosmos"),
+      });
+      const aminoMsg = aminoTypes.toAmino({
+        typeUrl: "/cosmos.staking.v1beta1.MsgUndelegate",
+        value: msg,
+      });
+      const expected: AminoMsgUndelegate = {
+        type: "cosmos-sdk/MsgUndelegate",
+        value: {
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          amount: coin(1234, "ucosm"),
+        },
+      };
+      expect(aminoMsg).toEqual(expected);
+    });
+  });
+  describe("fromAmino", () => {
+    it("works for MsgBeginRedelegate", () => {
+      const aminoMsg: AminoMsgBeginRedelegate = {
+        type: "cosmos-sdk/MsgBeginRedelegate",
+        value: {
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_src_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          validator_dst_address: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
+          amount: coin(1234, "ucosm"),
+        },
+      };
+      const msg = new AminoTypes({ ...createStakingAminoConverters("cosmos") }).fromAmino(aminoMsg);
+      const expectedValue: MsgBeginRedelegate = {
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorSrcAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        validatorDstAddress: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
+        amount: coin(1234, "ucosm"),
+      };
+      expect(msg).toEqual({
+        typeUrl: "/cosmos.staking.v1beta1.MsgBeginRedelegate",
+        value: expectedValue,
+      });
+    });
+
+    it("works for MsgCreateValidator", () => {
+      const aminoMsg: AminoMsgCreateValidator = {
+        type: "cosmos-sdk/MsgCreateValidator",
+        value: {
+          description: {
+            moniker: "validator",
+            identity: "me",
+            website: "valid.com",
+            security_contact: "Hamburglar",
+            details: "...",
+          },
+          commission: {
+            rate: "0.2",
+            max_rate: "0.3",
+            max_change_rate: "0.1",
+          },
+          min_self_delegation: "123",
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          pubkey: encodeBech32Pubkey(
+            { type: "tendermint/PubKeySecp256k1", value: "A08EGB7ro1ORuFhjOnZcSgwYlpe0DSFjVNUIkNNQxwKQ" },
+            "cosmos",
+          ),
+          value: coin(1234, "ucosm"),
+        },
+      };
+      const msg = new AminoTypes({ ...createStakingAminoConverters("cosmos") }).fromAmino(aminoMsg);
+      const expectedValue: MsgCreateValidator = {
+        description: {
+          moniker: "validator",
+          identity: "me",
+          website: "valid.com",
+          securityContact: "Hamburglar",
+          details: "...",
+        },
+        commission: {
+          rate: "0.2",
+          maxRate: "0.3",
+          maxChangeRate: "0.1",
+        },
+        minSelfDelegation: "123",
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        pubkey: {
+          typeUrl: "/cosmos.crypto.secp256k1.PubKey",
+          value: fromBase64("A08EGB7ro1ORuFhjOnZcSgwYlpe0DSFjVNUIkNNQxwKQ"),
+        },
+        value: coin(1234, "ucosm"),
+      };
+      expect(msg).toEqual({
+        typeUrl: "/cosmos.staking.v1beta1.MsgCreateValidator",
+        value: expectedValue,
+      });
+    });
+
+    it("works for MsgDelegate", () => {
+      const aminoMsg: AminoMsgDelegate = {
+        type: "cosmos-sdk/MsgDelegate",
+        value: {
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          amount: coin(1234, "ucosm"),
+        },
+      };
+      const msg = new AminoTypes({ ...createStakingAminoConverters("cosmos") }).fromAmino(aminoMsg);
+      const expectedValue: MsgDelegate = {
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        amount: coin(1234, "ucosm"),
+      };
+      expect(msg).toEqual({
+        typeUrl: "/cosmos.staking.v1beta1.MsgDelegate",
+        value: expectedValue,
+      });
+    });
+
+    it("works for MsgEditValidator", () => {
+      const aminoMsg: AminoMsgEditValidator = {
+        type: "cosmos-sdk/MsgEditValidator",
+        value: {
+          description: {
+            moniker: "validator",
+            identity: "me",
+            website: "valid.com",
+            security_contact: "Hamburglar",
+            details: "...",
+          },
+          commission_rate: "0.2",
+          min_self_delegation: "123",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        },
+      };
+      const msg = new AminoTypes({ ...createStakingAminoConverters("cosmos") }).fromAmino(aminoMsg);
+      const expectedValue: MsgEditValidator = {
+        description: {
+          moniker: "validator",
+          identity: "me",
+          website: "valid.com",
+          securityContact: "Hamburglar",
+          details: "...",
+        },
+        commissionRate: "0.2",
+        minSelfDelegation: "123",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+      };
+      expect(msg).toEqual({
+        typeUrl: "/cosmos.staking.v1beta1.MsgEditValidator",
+        value: expectedValue,
+      });
+    });
+
+    it("works for MsgUndelegate", () => {
+      const aminoMsg: AminoMsgUndelegate = {
+        type: "cosmos-sdk/MsgUndelegate",
+        value: {
+          delegator_address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+          amount: coin(1234, "ucosm"),
+        },
+      };
+      const msg = new AminoTypes({ ...createStakingAminoConverters("cosmos") }).fromAmino(aminoMsg);
+      const expectedValue: MsgUndelegate = {
+        delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
+        amount: coin(1234, "ucosm"),
+      };
+      expect(msg).toEqual({
+        typeUrl: "/cosmos.staking.v1beta1.MsgUndelegate",
+        value: expectedValue,
+      });
+    });
+  });
+});
